Access matchers: capture the matching map
Before, we tried to differentiate maps by using the out tuple id, but it never
worked because we were taking it after the projection operation that drops
tuples ids.  It also insufficient as maps with equal out tuple ids (and
assuming equal input space) may still have different output dimensionality.
Store the map instead.  Arguably, it is sufficient to store the space of the
map rather than the map itself.  However, it looks easier to get the map from
the matcher rather than look it up by space in the union map submitted to the
matcher.  Because of CoW, storing a copy of isl::map or isl::space is equally
cheap.