Access matchers: introduce folds
Two placeholders are folded if they are required to match the same access
expression potentially in different maps.  We represent them using a structure
similar to those used for disjoint sets: we maintain a vector of the same size
as the number of placeholders that indicates the first placeholder of the fold
for each of them.  By default, each placeholder is in folded only with itself.
Folded placeholders are not subject to the uniqueness-of-matches check.