Access matchers: draft API
Provide a preliminary API to build access matchers.  User should only construct
named placeholders.  They are constructed explicitly because they need an
isl::ctx.  Using the same named placeholder in multiple locations corresponds
to creating several (unnamed) placeholders that are folded together.  Placing
mulitple placeholders in a call to access() makes them belong to the same
group, multiple calls to access() form different groups.

Currently, from the user perspective, a placeholder is a replacement for a
single dimension of the input space because the latter is expected to be shared
between all access relations (e.g., the schedule space).

Extensions to consider:
- positional arguments instead of dim()
- negative python-like indexing for dim()
- fully wildcard placeholders and differentiating those from actual placeholders
- custom properties of access relations or individual dimensions as lambdas
- looking for specific dimensions rather than placeholders
- looking for constants
- matching the output space
- matching with different input spaces