access placeholders: introduce unique ids
Previously, test_access made use of a special class NamedPlaceholder to easier
construction of placeholder groups and folds (copies of a Placeholder instance
with the same name are folded).  Introduce unique ids into the Placeholder
class and use it directly in the tests.  This is not only useful for the future
API but also required to implement access transformers.