access matchers: rename makePS to allOf and move it to the header
The original name of a static function first introduced in the test implied it
was constructing a PlaceholderSet.  From the user perspective, there is no need
to know what a PlaceholderSet is, just that all the access patterns passed as
arguments should be matched, hence the new name.  Ideally, the type of value
returned by allOf would be unspecified, but it's not possible in C++.  In
future, make sure the user cannot construct or modify a PlaceholderSet
otherwise than by using the API provided.