access matchers: return matches as a placeholder-indexed object
The call to match() returns a list of matches.  In the current implemenetation,
an individual match is represented by a vector of per-dimension candidates
picked for each placeholder, coindexed with the vector of placeholders.
However, this vector is only supposed to be visible internally and is
constructed under the hood by allOf().  Instead, make individual match object
indexable by the placeholder, relying on the recently introduced id.