access matchers: factor out the payload from DimCandidate
Make DimCandidate and all classes that contain it class templates parameterized
by the type of payload they carry.  In the case of simple affine functions,
this payload previously existed as an integer position of the input dimension
that appeared in the affine function for the given dimension.  Reify it as
SingleInputDim class.  Make most matching functions function templates and move
them to access-inl.h from access.cc.  For functions with behavior specific to
matching simple affine expressions, require DimCandidate and likes to be
parameterized with SingleInputDim.  In the upcoming commit, these functions
will be abstracted away from the matching process.