make candidate selection and per-dim transfomation part of the payload
Move the implementations of the functions that find suitable candidates and
perform per-dimensional transformations to the static member functions of the
class used as candidate payload.  Provide simple template wrapper functions to
minimize changes in other places.  The current interface of these methods is
not very practical because it requires them to manipulate Placeholder
instances.  This will be addressed in future commits after more refactoring.

Filter customization for individual candidates is now possible by defining new
classes with different implementations of appendToCandidateList.