access machers: factor out the payload from Placeholder
Add another template argument to (Unpositioned)Placeholder and all classes that
contain it.  This class corresponds to the pattern the placeholder should match
(unlike candidates, pattern does not change throughout the match).  Reify the
existing pattern
  coefficient * #candidate + constant
into a new class SimpleAff and use it in test cases.  The candidate payload is
made aware of pattern payloads and only provides functions for matching and
replacement for those pattern payloads it corresponds to.