first attempt at stride matchers
Use the new separated payload matching infrastructure to match for strides.
Implement stride candidate and stirde pattern payload classes.  The former is
empty.  It always compares equals to other instances because we don't want the
placeholders to get reused in this case.  The latter contains the expected
stride value.  The stride is computed by computing the deltas along the last
output dimension of the access relation after projecting out the previous
dimensions.  (Note that access relations where output dimensions are
interdepenent in a way that does not allow to treat them per-dimension are not
representable as C-style arrays).  Only constant non-parametric strides are
supported.

Arguably, we may want candidates to always compare different to each other to
allow the caller to write the following
  auto x = stride(ctx, 1);
  allOf(access(x, x), access(x, x));