access matchers: don't group placeholders by dimension
The initial implementation was grouping placeholders by dimension before
looking for candidates.  This was a mostly premature optimization tied to the
mathching 1d affine expressions without constants, where placeholders could
reuse the candidate lists.  After the subsequent extensions, the only element
that was effectively reused per dimension was the projection of the access
relation on that dimension.  Projections are cheap so there is no need to cache
it.  Remove the grouping by dimension, and simplify the code around it.  This
will enable the upcoming support for negative dimension indexing and,
eventually, placeholders for more than one dimension.