access matchers: extract out per-dimension placeholders from the engine
The original implementation of access matchers used placeholders tied to a
particular dimension in the output space of the access relation.  A later
evolution introduced the separation between positioned and unpositioned
placeholders where only positioned placeholders can be used inside the matching
engine.  In some usecases, it may be necessary to have placeholders for entire
relations rather than individual output dimensions thereof.  For example, to
check that the relation is can be reconstructed from projections on individual
dimensions of its output space to start with.  Move the per-output-dimension
logic out of the matching engine into a newly introduced class template
FixedOutDimPattern.  Instantiations of this class template add the position of
the dimension in the output space into the set of pattern properties that all
candidates must exhibit.

This commit does not remove the separation between positioned and unpositioned
placeholder although the former becomes redundant to minimize the code changes.
It will be done in an upcoming commit.