access matchers: remove UnpositionedPlaceholder
The previous commit moved the logic of per-output-dimension placeholders to
their payloads, making the Positioned/Unpositioned placeholder distinction
redundant.  Remove UnpositionedPlaceholder and treat all placeholders as
unpositioned inside the engine (the pattern payload now takes care of
positioning).

For the sake of maintaining static type safety, introduce UnfixedOutDimPattern
class template to wrap payloads that should only appear wrapped into
FixedOutDimPattern when passed to the engine.  Overload access(...) function to
perform the transformations of the patterns when passed UnfixedOutDimPattern.
Provide the function template pattern_cast() to statically cast pattern
payloads of placeholder without recreating them (and thus keeping their ids) as
a means to support this.  Similarly to the original implementation, it remains
syntactically forbidden to mix UnfixedOutDimPattern and FixedOutDimPattern
instantiations in the same call to access(...) or allOf(...).