access matches: return all spaces in which placeholder appears
The initial implementation of Match class reportnig the match results was only
finding the first candidate assignment for the placeholder ID.  Therefore, it
was impossible to extract other spaces matched by folded placeholders, which
have the same ID by definition.  Introduce a new class MatchCandidate
containing a vector of candidate spaces assigned to the placeholder and use it
to report the result of matching.