tree builders: lazily construct node properties
Instead of storing the actual properties of the nodes to be constructed when
creating the builder, store function objects that can be used to construct
these properties.  This choice is motivated by the declarative lazy-evaluation
API around schedule tree builders: when a builder is constructed, it may serve
as a template for multiple trees, and the data members for these trees may not
exist yet.  This is, for example, the case when builders are used to
reconstruct (sub)trees captured by schedule tree matchers in an iterative
fashion.