tree builders: lazily evaluate subtrees
The original implementation of the subtree(...) builder constructor immediately
traversed the subtree rooted at the given schedule node.  This behavior was
inconsistent with the other builder constructors that were taking function
objects only evaluated during the actual tree building.  Store a callback
constructing a subtree builder for leaf builders and call it when building the
tree.  Take this callback in subtree builder constructor.  Provide a
convenience overload for schedule tree nodes.  Also expose the function for
eagerly constructing a builder from a subtree.