[CircleCI]: cache isl and pet installations
isl and pet submodules are expected to change rarely, but they take up to 90%
of the build time.  Cache them in CircleCI instead of rebuilding each time.

In particular, cache the install-isl and install-pet directories along
with the special "id" file containing commit hashes of both isl and pet.  Load
from cache before attempting to build isl and pet, generate the new "id" file
and only build pet if the stored and the generated files differ.  Note that
this still requires to checkout submodules, but the checkout duration is
negligeable.