tree builders: capture coincidence and permutability of bands
Originally, only the partial schedule of a band was set up by the builder.
This is not sufficient in multiple practical cases where schedule trees must be
transformed to expose, e.g., parallelism.  Introduce a simple data-wrapper
class stroring partial schedules along with coincidence and permutability
flags.  Use instances of this class instead of multi_union_pw_aff to construct
bands in builders.  Make sure information is propagated when building subtrees.

This functionality is tested using recently introduced functions to merge
nested bands if they are permutable and to set the corresponding flag.