Parser::getScop: take isl ctx instead of allocating a new one
The original implementation of the getScop method was allocating a new isl ctx
every time it was called.  This is dangerous if the caller attempts to mix isl
objects allocated in different contexts, for example those that  originated
from repeated parses of the same input file.  In particular, one test case was
also allocating a separate context.  Take the context as argument of getScop
instead.  Expose a function that allocates a C++-wrapped context with pet
options so that the callers of Parser don't have to include pet.h.  Allow
ScopedCtx to claim ownership over an existing context and free it on leaving
the scope.  This cleans up memory leaks due to not freeing the context after
using the parser.