access matcher: move group validity checks to PlaceholderSet
Instead of hardcoding the validity checks of candidate-to-placeholder
assignments with respect to placeholder groups and folds in the "match"
function, put them into a member function of PlaceholderSet.  This opens the
way for specifying group/fold behavior when an access matcher is constructed.
It is not yet possible because the type of PlaceholderSet is hardcoded in the
match() call.  This also simplifies the code by removing template arguments and
lambdas.