access matcher: make PlaceholderSet type a template parameter of match()
This allows the caller to redefine group/fold validation behavior (e.g. to
allow different placeholders from different folds to match the the same
candidate) when calling the matcher by providing a different class implementing
behavior compatible with PlaceholderSet (can't wait for C++20 concepts!) and a
with different isSuitableCombination() behavior.