Introduce an API for cosntructing PlaceholderGroupedSet
Overload access(...) and allOf(...) to take a special type ArrayPlaceholder to
identify accesses to the same or different arrays.  The mechanism is similar to
placeholder folds and was introduced in the previous commits.  ArrayPlaceholder
is a specific named instantiation of Placeholder with "char" as pattern and
candidate payloads since they are not used.

Note that, unlike regular placeholders, matched array ids cannot be extracted
from the match results easily.  The caller must look at the spaces of matching
relations and extract the array id from there.  Fixing this will require to
move the matched space from the matcher engine to the candidate payload, which
may have further repercussions on the transformation engine.  At the same, time
this will make access matcher engine extremely flexible, to the point that it
can be considered generic relation matcher engine.