subtree builders: explicitly specify type for nodes with no payload
The original implementation of subtree builder creation had been explicitly
copying the type of the node to the builder.  The update in
9bf611fbea190d5091a9e220b9d1dfde7497b6c2 removed this copy and chose to rely on
public API for builder construction instead.  However, for nodes with no
payload, public API functions are not called, leaving the type field of the
builder unset.  This has led to builders being in an incorrect state when nodes
with no payload are encountered in subtree builder construction.  Copy the type
explicitly in such cases.