pet codegen: store the state of AST build instead of ref2expr
Instead of computing and storing the mapping between reference identifiers and
AST expressions while building the AST itself, store the state of the AST build
itself.  Build the mapping when actually generating code from the AST.  This is
potentially expensive in terms of used memory (although I expect isl
copy-on-write semantics to help with the storage of duplicates), but localizes
the calls to pet infrastructure in a single callback instead of two.  In an
upcoming commit, the codegen API will be extended to accept a codegen callback
for injected statements that should not go through the pet infrastructure.
From the API design point of view, it is better to localize the codegen for the
injected code to a single function.  This choice may be reconsidered in the
future.