First attempt at implementing Polly matmul optimiztion.
The optimization performed is based on the paper: High-Performance Generalized
Tensor Operations: A Compiler-Oriented Approach. The algorithm implemented is
at page 7 of the mentioned paper. The missing transformation is the packing of
arrays A and B. For tile size we assume manually set values, in this particular
case (32, 32, 32) for the macro-kernel and (2, 2, 1) for the micro-kernel.
Notice that there is a mismatch between what Polly does and the optimization
steps described in the paper. This should be clarified.